+++ progress.py
@@ -47,12 +47,12 @@ class ThreadProgress():
                 (("status_code" in result and result["status_code"] > 399) or\
                  ("success" in result and not result["success"])):
                 
-                print (message.SEPRATE.format(util.format_error_message(result)))
-                if self.open_console: util.show_panel()
+                if self.open_console: 
+                    util.show_output_panel(message.SEPRATE.format(util.format_error_message(result)))
 
                 settings = context.get_toolingapi_settings()
                 delay_seconds = settings["delay_seconds_for_hidden_console"]
-                sublime.set_timeout_async(util.hide_panel, delay_seconds * 1000)
+                sublime.set_timeout_async(util.hide_output_panel, delay_seconds * 1000)
             else:
                 sublime.status_message(self.success_message)
             return